app-engine/GCP-App-Engine-flexible/python-flask-app/requirements.txt
Flask==2.1.0;python_version > '3.6'
google-cloud-bigquery==2.34.4
gunicorn==20.1.0